requirements.txt
fastapi
uvicorn
jinja2
langchain
langchain_openai
langchain_core
python-multipart
langchain[groq]
python-dotenv